Improved a lot the example project. Texture2D is no longer supported for blocks, use BlockTexture2D instead.

+++ MineCloneProject/src/MCP/Maths/Maths.h
@@ -63,7 +63,7 @@ namespace MC
 		 return ((start * (1.0f - alpha)) + (end * alpha));	
 	}
 
-	inline int floorf(const float val)
+	inline float floorf(const float val)
 	{
 		return ::floorf(val);
 	}
